Moved testing classes to separate file
