+++ dev-requirements.txt
@@ -2,7 +2,6 @@
 flake8==4.0.1
 black==21.12b0
 websocket-client==1.4.2
-maturin==0.12.11
 isort==5.10.1
 pre-commit==2.19.0
 commitizen==2.27.1
